Scaffolded unit tests
