+++ tests/WPStrava/SOMEnglishTest.php
@@ -1,0 +1,11 @@
+<?php
+
+use \WP_Mock\Tools\TestCase;
+
+class WPStrava_SOMEnglishTest extends TestCase {
+
+	public function test_true() {
+		$som = new WPStrava_SOMEnglish();
+		$this->assertInstanceOf( 'WPStrava_SOMEnglish', $som );
+	}
+}
